feat: Add filterData class
