+++ src/types/Loader.ts
@@ -12,30 +12,3 @@ export interface JsonProducts {
     skip: number;
     limit: number;
 }
-
-export enum FiltersType {
-    category = 'category',
-    brand = 'brand',
-    price = 'price',
-    stock = 'stock',
-}
-
-export enum SortType {
-    rating = 'rating',
-    title = 'title',
-}
-
-export type MaxMin = {
-    min: number;
-    max: number;
-};
-
-export interface FilterCollection {
-    type: FiltersType;
-    keys: string[] | MaxMin;
-}
-
-export enum SortDirection {
-    up,
-    down,
-}
